api-back/requirements.txt
asgiref==3.8.0
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
icalendar==5.0.12
idna==3.6
importlib-resources==6.4.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
pillow==10.3.0
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
zipp==3.18.1
python-telegram-bot==20.0
